Dispatch basic utility methods to helpers.utility
