+++ src/chart.js
@@ -5,9 +5,6 @@ var Chart = require('./core/core')();
 
 Chart.helpers = require('./helpers/index');
 
-// @todo dispatch these helpers into appropriated helpers/helpers.* file and write unit tests!
-require('./core/core.helpers')(Chart);
-
 Chart.Animation = require('./core/core.animation');
 Chart.animationService = require('./core/core.animations');
 Chart.controllers = require('./controllers/index');
